port_monitoring/requirements.txt
numpy
scikit-learn
psutil
joblib
pyyaml
